requirements.txt
{\rtf1\ansi\ansicpg1252\cocoartf2822
\cocoatextscaling0\cocoaplatform0{\fonttbl\f0\fnil\fcharset0 .AppleSystemUIFontMonospaced-Regular;}
{\colortbl;\red255\green255\blue255;\red39\green38\blue34;\red42\green40\blue35;}
{\*\expandedcolortbl;;\cssrgb\c20392\c19608\c17647;\cssrgb\c21569\c20784\c18431\c3922;}
\paperw11900\paperh16840\margl1440\margr1440\vieww11520\viewh8400\viewkind0
\deftab720
\pard\pardeftab720\partightenfactor0

\f0\fs18\fsmilli9188 \cf2 \cb3 \expnd0\expndtw0\kerning0
\outl0\strokewidth0 \strokec2 blinker==1.9.0\
certifi==2025.7.14\
charset-normalizer==3.4.2\
click==8.2.1\
Flask==3.1.1\
flask-cors==6.0.0\
Flask-SQLAlchemy==3.1.1\
greenlet==3.2.3\
idna==3.10\
itsdangerous==2.2.0\
Jinja2==3.1.6\
MarkupSafe==3.0.2\
pillow==11.3.0\
requests==2.32.4\
SQLAlchemy==2.0.41\
typing_extensions==4.14.0\
urllib3==2.5.0\
Werkzeug==3.1.3
\fs21 \
}